Use string replace to replace inline markup with rendered templates
Previously the inline markup was replaced with the rendered template
using BeautifulSoup's replaceWith() method.

The result of render_to_string() is, as you might expect, a string. The
problem is that, in current versions of BeautifulSoup, if replaceWith()
is passed a string, it get turns into a BeautifulSoup NavigableString.
Any HTML characters -- notably '<' and '>' -- contained in Navigable
Strings are escaped. This results in an unusable template.

To get around this issue, the parsed soup is now turned back into a
string. Any inline markup that BeautifulSoup finds is turned into a
string. Python's str.replace() is then used to replace any instance of
the inline markup string with the rendered template string in the parsed soup
string. (Whoo, that's a lot of strings.) Finally, the parsed soup string is returned.

Treating the parsed soup as a string does not seem like the best way to
deal with this (why even parse it in the first place?). It seems to work
in my tests, but I'm not sure if there will be any unintended
consequences.

I'm open to better solutions.