All inline attributes are now added to the context
The attributes "id" (or "ids") and "type" are required in the inline
markup. Any other attribute can be added to the markup, but previously
only the "class" attribute was extracted and added to the template
context. Now, all attributes are added to the context.

This can potentially add a lot of flexibility, depending on your
templates. If you are including a photo, you might have previously
done this:

    <inline type="media.photo" id="1" class="small_left" />

Now, you could do this:

    <inline type="media.photo" id="1" size="small" align="left" />

Or maybe you want to add a caption to the photo:

    <inline type="media.photo" id="1" size="small" align="left"
    caption="Sweet photo, dude." />

Easy!

There are two backwards-incompatible changes to note:

 1. Previously, the context had an entry named "content_type" with the
 value of "app.label". This entry has been replaced with one named
 "type" (which is consistent with the actual inline markup). The value
 is identical.
 2. The Django settings were previously added the context. I don't see
 any reason for this, so they're no longer added.