support configurable object managers
By specifying a manager for an inline type you can restrict the objects
that are rendered.

For example, you may have a `published` manager on a `Photo` model in a
`media` app which filters the queryset such that only published photos
are returned. You could configure that manager to be used in place of
the default `all` manager:

    INLINES_MANAGERS = {
        'media.photo': 'published',
    }

This would result in calling `Photo.objects.published().filter(...)`
anytime an inline tag with `type="media.photo"` is encountered.